app/requirements.txt
Flask
SQLAlchemy
Pyrebase4
alembic
Flask_Migrate
firebase_admin
Flask_SQLAlchemy
beautifulsoup4
Pillow
flask-msearch
python-dotenv